% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFilename.R
\name{addFilename}
\alias{addFilename}
\title{Add column to data containing name of file}
\usage{
addFilename()
}
\value{
A `FieldRef` Expression that refers to the filename augmented column.
}
\description{
In arrow data retrieval, add file name column to tables. Replicated from https://github.com/apache/arrow/blob/main/r/R/dplyr-funcs-augmented.R because add_filename() is unexported in arrow.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
