% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignSpCols.R
\name{alignSpCols}
\alias{alignSpCols}
\title{Rename column names in old sensor position files}
\usage{
alignSpCols(sptable)
}
\arguments{
\item{sptable}{A data frame of sensor positions}
}
\value{
A data frame with updated naming
}
\description{
Rename columns in old sensor positions files to match current names
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
