% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignClasses.R
\name{assignClasses}
\alias{assignClasses}
\title{Assign correct column classes}
\usage{
assignClasses(dt, inVars)
}
\arguments{
\item{dt}{A data frame}

\item{inVars}{The variables expected in the df}
}
\value{
A data frame with corrected column classes
}
\description{
Use the variables file to assign classes to each column in each data file
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
