% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byPointsAOP.R
\name{byPointsAOP}
\alias{byPointsAOP}
\title{Serially download a set of AOP files for a given set of woody vegetatation structure points.}
\usage{
byPointsAOP(dpID, site = "SJER", year = "2017", check.size = TRUE,
  savepath = NA, allSites = F)
}
\arguments{
\item{dpID}{The identifier of the NEON data product to pull, in the form DPL.PRNUM.REV, e.g. DP1.10023.001}

\item{site}{The four-letter code of a single NEON site, e.g. 'CLBJ'.}

\item{year}{The four-digit year to search for data. Defaults to 2017.}

\item{check.size}{T or F, should the user be told the total file size before downloading? Defaults to T. When working in batch mode, or other non-interactive workflow, use check.size=F.}

\item{savepath}{The file path to download to. Defaults to NA, in which case the working directory is used.}

\item{savepath}{The file path to download to. Defaults to NA, in which case the working directory is used.}
}
\value{
A folder in the working directory, containing all files meeting query criteria.
}
\description{
Query the API for AOP data by site, year, and product, and download all files found, preserving original
folder structure. Downloads serially to avoid overload; may take a very long time.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Ben Weinstein \email{ben.weinstein@weecology.org}
}
