% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUrlExp.R
\name{checkUrlExp}
\alias{checkUrlExp}
\title{Check for expiration of a signed url}
\usage{
checkUrlExp(urlIn)
}
\arguments{
\item{urlIn}{A signed url to be checked}
}
\value{
Expiration date-time for the url
}
\description{
For a signed url, check whether the expiration date-time has been reached
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
