% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVarFields.R
\name{checkVarFields}
\alias{checkVarFields}
\title{Check for differences in field names among variables files}
\usage{
checkVarFields(variableSet, tableName)
}
\arguments{
\item{variableSet}{A list of file paths or urls to variables files}

\item{tableName}{Name of table to check for differences}
}
\value{
TRUE or FALSE: were there any mismatches in field names and data types among the files?
}
\description{
For a set of variables files, check whether there are any differences in the set of field names and data types for a particular table
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
