% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_bundles.R
\docType{data}
\name{chem_bundles}
\alias{chem_bundles}
\title{Bundled chemistry data product information}
\format{
A data frame with 2 variables:
\describe{
  \item{product}{Data product ID of a terrestrial chemistry product}
  \item{homeProduct}{Data product ID of the corresponding home data product}
}
}
\source{
NEON data product bundles
}
\usage{
chem_bundles
}
\description{
A dataset containing NEON data product codes of terrestrial chemistry data products and the "home" data products they are bundled with.
}
\keyword{datasets}
