% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUp.R
\name{cleanUp}
\alias{cleanUp}
\title{Clean up folder after stacking}
\usage{
cleanUp(folder, orig)
}
\arguments{
\item{folder}{The file path to the folder that needs to be cleaned up (the root directory of the data package)}

\item{orig}{The list of files that were present in the folder before unzipping and stacking}
}
\value{
Only the folders created during unzip will be deleted. All custom folders/files and the stackedFiles output folder will be retained.
}
\description{
Remove unzipped monthly data folders
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
