% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convByteSize.R
\name{convByteSize}
\alias{convByteSize}
\title{Convert a number of bytes into megabytes or gigabytes}
\usage{
convByteSize(objSize)
}
\arguments{
\item{objSize}{The size in bytes}
}
\value{
The size of the file in megabytes or gigabytes
}
\description{
For any number of bytes, convert to a number of MB or GB
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
