% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateConvert.R
\name{dateConvert}
\alias{dateConvert}
\title{Convert date stamps from character}
\usage{
dateConvert(dates, useFasttime = FALSE)
}
\arguments{
\item{dates}{A vector of date values in character format [character]}

\item{useFasttime}{Should the fasttime package be used for date conversion? Defaults to false. [logical]}
}
\value{
A POSIXct vector, if possible; if conversion was unsuccessful, the original vector is returned
}
\description{
Attempt to convert date stamps from character, iterating through known NEON date formats
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
