% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNEONFile.R
\name{downloadNEONFile}
\alias{downloadNEONFile}
\title{Download a file from the NEON API}
\usage{
downloadNEONFile(url, outpath, useragent, token)
}
\arguments{
\item{url}{The url pointing to the file to be downloaded}

\item{outpath}{The file path to save the file to}

\item{useragent}{The user agent to send to the API}

\item{token}{The user's API token}
}
\value{
A status message indicating whether the download was successful or not
}
\description{
Download a file via the NEON API, using a token as necessary.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
