% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDatatables.R
\name{findDatatables}
\alias{findDatatables}
\title{Find data tables}
\usage{
findDatatables(folder, fnames = T)
}
\arguments{
\item{folder}{The folder of the outputs}

\item{fnames}{Full names - if true, then return the full file names including enclosing folders, if false, return only the file names}
}
\value{
a data frame of file names
}
\description{
List the names of the data tables within each folder
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
