% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTablesByFormat.R
\name{findTablesByFormat}
\alias{findTablesByFormat}
\title{Find unique data tables in dataset}
\usage{
findTablesByFormat(datatables)
}
\arguments{
\item{datatables}{A list of data files}
}
\value{
An array of unique table names and their types
}
\description{
Find the unique data tables that are present in the dataset (e.g., 2 minute vs 30 minute, or pinning vs identification data) and their types, based on the file name formatting. Adapted from findTablesUnique().
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
