% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatReadme.R
\name{formatReadme}
\alias{formatReadme}
\title{Get a genericized version of a readme file by removing info that is specific to a site-month or data query.}
\usage{
formatReadme(savepath, dpID)
}
\arguments{
\item{savepath}{A data frame containing the readme contents.}

\item{dpID}{The data product identifier}
}
\description{
Read in a complete readme file, and return a genericized file.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
