% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPI.R
\name{getAPI}
\alias{getAPI}
\title{Get the data from API}
\usage{
getAPI(apiURL, token = NA_character_)
}
\arguments{
\item{apiURL}{The API endpoint URL}

\item{token}{User specific API token (generated within data.neonscience.org user accounts). Optional.}
}
\description{
Accesses the API with options to use the user-specific API token generated within data.neonscience.org user accounts.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Nate Mietkiewicz \email{mietkiewicz@battelleecology.org}
}
\keyword{internal}
