% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIHeaders.R
\name{getAPIHeaders}
\alias{getAPIHeaders}
\title{Get only headers from API}
\usage{
getAPIHeaders(apiURL, token = NA_character_)
}
\arguments{
\item{apiURL}{The API endpoint URL}

\item{token}{User specific API token (generated within neon.datascience user accounts). Optional.}
}
\description{
Accesses the API with options to use the user-specific API token generated within neon.datascience user accounts.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
