% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttributes.R
\name{getAttributes}
\alias{getAttributes}
\title{Extract attributes from eddy covariance H5 files}
\usage{
getAttributes(fil, sit)
}
\arguments{
\item{fil}{File path to the H5 file to extract attributes from [character]}

\item{sit}{The site, for site attributes. Must match site of file path. [character]}
}
\value{
A data frame with one row containing the extracted attributes
}
\description{
Extract attribute metadata from H5 files
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
