% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCitation.R
\name{getCitation}
\alias{getCitation}
\title{Get a Bibtex citation for NEON data with a DOI}
\usage{
getCitation(doi)
}
\arguments{
\item{doi}{The DOI of the data to be cited [character]}
}
\value{
A character string containing the Bibtex citation
}
\description{
Use the DOI Foundation API to get Bibtex-formatted citations for NEON data. Technically this function will work for any DOI, but is not exported since there are much more thorough wrappers for DOI APIs.
}
\examples{
\dontrun{
# Get the citation for Breeding landbird point counts (DP1.10003.001), RELEASE-2023
cit <- getCitation(doi="10.48443/S730-DY13")
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
