% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEddyLog.R
\name{getEddyLog}
\alias{getEddyLog}
\title{Get the full issue log set for the SAE bundle}
\usage{
getEddyLog(token = NA_character_)
}
\arguments{
\item{token}{User specific API token (generated within data.neonscience.org user accounts)}
}
\value{
A table of issues reported for the data product.
}
\description{
Use the NEON API to get the issue log from all products in the bundle in a user-friendly format
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
