% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilesize.R
\name{getFilesize}
\alias{getFilesize}
\title{Get a file's size in megabytes}
\usage{
getFilesize(filepath)
}
\arguments{
\item{filepath}{The path to the file}
}
\value{
The size of the file in megabytes
}
\description{
For any file, read the size and translate into a human readable string
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
