% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIssueLog.R
\name{getIssueLog}
\alias{getIssueLog}
\title{Get the issue log for a specific data product}
\usage{
getIssueLog(dpID = NA, token = NA_character_)
}
\arguments{
\item{dpID}{The data product identifier, formatted as DP#.#####.###}

\item{token}{User specific API token (generated within data.neonscience.org user accounts)}
}
\value{
A table of issues reported for the data product.
}
\description{
Use the NEON API to get the issue log in a user-friendly format
}
\examples{
# Get documentation and availability of plant foliar properties data product
cfcIssues <- getIssueLog("DP1.10026.001")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
