% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeonDOI.R
\name{getNeonDOI}
\alias{getNeonDOI}
\title{Get either a list of NEON DOIs, or the DOI for a specific data product and release}
\usage{
getNeonDOI(dpID = NA_character_, release = NA_character_)
}
\arguments{
\item{dpID}{The data product identifier, formatted as DP#.#####.### [character]}

\item{release}{Name of a specific release, e.g. RELEASE-2022 [character]}
}
\value{
A table of data product IDs and DOIs.
}
\description{
Use the DataCite API to get NEON data DOIs in a user-friendly format
}
\examples{
\dontrun{
# Get all NEON data DOIs
allDOIs <- getNeonDOI()
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
