% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProductSensors.R
\name{getProductSensors}
\alias{getProductSensors}
\title{Get data product-sensor relationships}
\usage{
getProductSensors()
}
\value{
A data frame
}
\description{
Pull all data from the NEON API /products endpoint, create a data frame with data product ID, data product name, and sensor type.
}
\examples{
sensors <- getProductSensors()
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
