% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadmePublicationDate.R
\name{getReadmePublicationDate}
\alias{getReadmePublicationDate}
\title{Scrape the publication date from each ReadMe file}
\usage{
getReadmePublicationDate(savepath, out_filepath)
}
\arguments{
\item{savepath}{The root folder directory where the ReadMe files are located.}

\item{out_filepath_name}{The output directory and filename.}
}
\description{
Given a directory, this will recursively list all of the ReadMe files that were unzipped.
This should result in a single text file with a list of all of the publication dates from the ReadMe file.
}
\examples{
\dontrun{
# To unzip and merge files downloaded from the NEON Data Portal
stackByTable("~/NEON_par.zip")

# To unzip and merge files downloaded using zipsByProduct()
stackByTable("~/filesToStack00024", folder=T)
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Nathan Mietkiewicz \email{mietkiewicz@battelleecology.org}
}
