% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecentPublication.R
\name{getRecentPublication}
\alias{getRecentPublication}
\title{Returns the most recent files for those that do not need stacking}
\usage{
getRecentPublication(inList)
}
\arguments{
\item{inList}{The list of files.}
}
\description{
Given a list of files, this will order and return the file with the most recent publication date.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Nathan Mietkiewicz \email{mietkiewicz@battelleecology.org}
}
