% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTitle.R
\name{getTitle}
\alias{getTitle}
\title{Get NEON data product title}
\usage{
getTitle(filename)
}
\arguments{
\item{filename}{A NEON file name}
}
\value{
A title for the respective GeoCSV file
}
\description{
Create a title for a NEON data CSV file
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
