% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariablesEddy.R
\name{getVariablesEddy}
\alias{getVariablesEddy}
\title{Get variable names and units from SAE H5 files}
\usage{
getVariablesEddy(tabList)
}
\arguments{
\item{tabList}{A list of SAE data tables}
}
\value{
A table of variable names and units, aggregated from the input tables
}
\description{
Extract variable names and units from SAE H5 files and return in user-friendly form. Used in stackEddy(), not intended for independent use.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
