% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarsEC.R
\name{getVarsEC}
\alias{getVarsEC}
\title{Extract list of eddy covariance variables from HDF5 files}
\usage{
getVarsEC(filepath)
}
\arguments{
\item{filepath}{The folder containing the H5 files [character]}
}
\value{
A data frame of the specified variables
}
\description{
Extracts a list of variables contained in a set of HDF5 files. Specific to eddy covariance data product: DP4.00200.001
}
\examples{
\dontrun{
# read variables from a file in the working directory
getVarsEC(filepath=getwd())
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
