% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZipUrls.R
\name{getZipUrls}
\alias{getZipUrls}
\title{Get and store the file names, S3 URLs, file size, and download status (default = 0) in a data frame}
\usage{
getZipUrls(
  tmp.files,
  month.urls = month.urls,
  avg,
  package,
  max.pub,
  max.pub.site,
  dpID
)
}
\arguments{
\item{tmp.files}{The monthly API URL for the AOP files}

\item{month.urls}{The monthly API URL for the AOP files}

\item{avg}{Global variable for averaging interval}

\item{package}{Global varaible for package type (basic or expanded)}

\item{max.pub}{Global variable for the most recent URL publication date}

\item{max.pub.site}{Global variable for the most recent site  from publication date}

\item{dpID}{Global variable for data product ID}
}
\value{
A dataframe comprised of file names, S3 URLs, file size, and download status (default = 0)
}
\description{
Used to generate a data frame of available zipfile URLs.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
Christine Laney \email{claney@battelleecology.org}
}
