% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePosColumns.R
\name{makePosColumns}
\alias{makePosColumns}
\title{Create position (horizontal and vertical) columns}
\usage{
makePosColumns(d, datafl)
}
\arguments{
\item{d}{A data table}

\item{datafl}{A data file name}
}
\value{
A data table with new columns
}
\description{
For instrumented meteorological data products, create position (horizontal and vertical) columns based on values
embedded in the file names.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
