% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_bundles.R
\docType{data}
\name{other_bundles}
\alias{other_bundles}
\title{Bundled vegetation and sediment data product information}
\format{
A data frame with 2 variables:
\describe{
  \item{product}{Data product ID of a product}
  \item{homeProduct}{Data product ID of the corresponding home data product}
}
}
\source{
NEON data product bundles
}
\usage{
other_bundles
}
\description{
A dataset containing NEON data product codes of vegetation and sediment data products and the "home" data products they are bundled with.
}
\keyword{datasets}
