% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelFileAOP.R
\name{ParallelFileAOP}
\alias{ParallelFileAOP}
\title{Parallel download all AOP files for a given site, year, and product}
\usage{
ParallelFileAOP(dpID, site = "SJER", year = "2017", check.size = TRUE,
  savepath = NA, cores = 1)
}
\arguments{
\item{dpID}{The identifier of the NEON data product to pull, in the form DPL.PRNUM.REV, e.g. DP1.10023.001}

\item{site}{The four-letter code of a single NEON site, e.g. 'CLBJ'.}

\item{year}{The four-digit year to search for data. Defaults to 2017.}

\item{check.size}{T or F, should the user be told the total file size before downloading? Defaults to T. When working in batch mode, or other non-interactive workflow, use check.size=F.}

\item{savepath}{The file path to download to. Defaults to NA, in which case the working directory is used.}

\item{cores}{Number of cores to parallel download}
}
\value{
A folder in the working directory, containing all files meeting query criteria.
}
\description{
Query the API for AOP data by site, year, and product, and download all files found, preserving original
folder structure. Downloads serially to avoid overload; may take a very long time.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
Christine Laney \email{claney@battelleecology.org}
Ben Weinstein \email{ben.weinstein@weecolog.org}
}
