% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/plots.R
\docType{data}
\name{plots}
\alias{plots}
\alias{plots}
\title{Spatial Location of NEON plots
All_Neon_TOS_Polygons_V5.shp}
\format{A sf object with 3756 rows and 37 variables:}
\source{
\url{http://data.neonscience.org/data-product-view?dpCode=DP1.10098.001}

\url{http://data.neonscience.org/data-product-view?dpCode=DP1.10098.001}
}
\usage{
plots

plots
}
\description{
Spatial Location of NEON plots
All_Neon_TOS_Polygons_V5.shp

All_Neon_TOS_Polygons_V5.shp
}
\keyword{datasets}
