% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTableNEON.R
\name{readTableNEON}
\alias{readTableNEON}
\title{Read a NEON data table with correct data types for each variable}
\usage{
readTableNEON(dataFile, varFile, useFasttime = FALSE)
}
\arguments{
\item{dataFile}{A data frame containing a NEON data table, or the filepath to a data table to load}

\item{varFile}{A data frame containing the corresponding NEON variables file, or the filepath to the variables file}

\item{useFasttime}{Should the fasttime package be used to read date-time variables? Defaults to false.}
}
\value{
A data frame of a NEON data table, with column classes assigned by data type
}
\description{
Load a table into R, assigning classes to each column based on data types in variables file; or convert a table already loaded
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
