% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDups.R
\name{removeDups}
\alias{removeDups}
\title{Remove duplicates from a data table based on a provided primary key; flag duplicates that can't be removed.}
\usage{
removeDups(data, variables, table)
}
\arguments{
\item{data}{A data frame containing data from a NEON data table}

\item{variables}{The variables file containing the data table}

\item{table}{The name of the table. Must match one of the table names in 'variables'}
}
\value{
A modified data frame with duplicates removed and a flag field added and populated.
}
\description{
NEON observational data may contain duplicates; this function removes exact duplicates and attempts to resolve non-exact duplicates, and flags duplicates that can't be resolved.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
