% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSrfDups.R
\name{removeSrfDups}
\alias{removeSrfDups}
\title{Aggregate science review flag files to unique records}
\usage{
removeSrfDups(srftable)
}
\arguments{
\item{srftable}{A data frame of science review flags}
}
\value{
A data frame of science review flags with duplicates removed
}
\description{
Aggregate repeated science review flags to a unique set matching the download
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
