% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemaAllStrings.R
\name{schemaAllStrings}
\alias{schemaAllStrings}
\title{Create an arrow schema with every variable coded as a string field.}
\usage{
schemaAllStrings(variables)
}
\arguments{
\item{variables}{A data frame containing a NEON variables file for a single table, or a set of field names.}
}
\value{
An arrow schema for the relevant data table with all variables set to string.
}
\description{
Use the field names in a NEON variables file to create an arrow schema of all strings, or, if no variables file is available, read the header and assign everything as string.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
