% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemaFromVarDuck.R
\name{schemaFromVarDuck}
\alias{schemaFromVarDuck}
\title{Create a duckdb schema from a NEON variables file.}
\usage{
schemaFromVarDuck(variables, tab, package)
}
\arguments{
\item{variables}{A data frame containing a NEON variables file, or a url pointing to a NEON variables file.}

\item{tab}{The name of the table to generate a schema from.}

\item{package}{Should the schema be created for the basic or expanded package?}
}
\value{
A duckdb schema for the relevant data table.
}
\description{
Use the field names and data types in a NEON variables file to create a duckdb schema.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
