% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackDataFilesParallel.R
\name{stackDataFilesParallel}
\alias{stackDataFilesParallel}
\title{Join data files in a unzipped NEON data package by table type}
\usage{
stackDataFilesParallel(folder, nCores = 1, forceParallel = FALSE)
}
\arguments{
\item{folder}{The location of the data}

\item{nCores}{The number of cores to parallelize the stacking procedure. By default it is set to a single core.}

\item{forceParallel}{If the data volume to be processed does not meet minimum requirements to run in parallel, this overrides. Set to FALSE as default.}
}
\value{
One file for each table type is created and written.
}
\description{
Given a folder of unzipped files (unzipped NEON data file), do a full join of all data files, grouped by table type.
This should result in a small number of large files.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
