% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackEC.R
\name{stackEC}
\alias{stackEC}
\title{Extract eddy covariance data from HDF5 format}
\usage{
stackEC(filepath, level = "dp04", var = NA, avg = NA)
}
\arguments{
\item{filepath}{The folder containing the H5 files [character]}

\item{level}{The level of data to extract; one of dp01, dp02, dp03, dp04, dp0p [character]}

\item{var}{The variable set to extract, e.g. co2Turb [character]}

\item{avg}{The averaging interval to extract, in minutes [numeric]}
}
\value{
A data frame of the specified variables
}
\description{
Convert data of choice from HDF5 to tabular format. Specific to eddy covariance data product: DP4.00200.001
}
\examples{
\dontrun{
# To extract and merge Level 4 data tables, where data files are in the working directory
flux <- stackEC(filepath=getwd(), level='dp04', var=NA, avg=NA)
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
