% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackFrameFiles.R
\name{stackFrameFiles}
\alias{stackFrameFiles}
\title{Stack data frame (per sample) files}
\usage{
stackFrameFiles(framefiles, dpID, seqType = NA_character_, cloud.mode = FALSE)
}
\arguments{
\item{framefiles}{A vector of file paths to data frame files}

\item{dpID}{Data product ID of the product to be stacked}

\item{seqType}{For microbe community files, 16S or ITS}

\item{cloud.mode}{Are data stacked in a cloud-to-cloud transfer?}
}
\value{
A data frame of the stacked version of the input tables
}
\description{
Stacking and re-naming workflow for data frame files, published as a data table for each sample
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
