% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeStampSet.R
\name{timeStampSet}
\alias{timeStampSet}
\title{Generate a consensus set of time stamps from a set of input tables.}
\usage{
timeStampSet(tabList)
}
\arguments{
\item{tabList}{A list of SAE data tables}
}
\value{
A table of time stamps (start and end times) aggregated from the input tables
}
\description{
Generate consensus SAE time stamps from a set of tables. Used in stackEddy(), not intended for independent use.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
