% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenCheck.R
\name{tokenCheck}
\alias{tokenCheck}
\title{Check for expired API token}
\usage{
tokenCheck(token)
}
\arguments{
\item{token}{User specific API token (generated within data.neonscience.org user accounts)}
}
\value{
Returns a token value: either the original token, if unexpired, or NA, if the token has expired
}
\description{
Extracts the expiration date from API token and check whether it has expired.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
