% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenDate.R
\name{tokenDate}
\alias{tokenDate}
\title{Get expiration date for a NEON API token}
\usage{
tokenDate(token)
}
\arguments{
\item{token}{User specific API token (generated within data.neonscience.org user accounts)}
}
\value{
Returns the date when the token will expire (or has expired).
}
\description{
Extracts the expiration date from a NEON API token.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
