% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipZipfile.R
\name{unzipZipfile}
\alias{unzipZipfile}
\title{Unzip a zip file either at just the top level or recursively through the file}
\usage{
unzipZipfile(zippath, outpath = substr(zippath, 1, nchar(zippath) - 4),
  level = "all")
}
\arguments{
\item{zippath}{The filepath of the input file}

\item{outpath}{The name of the folder to save unpacked files to}

\item{level}{Whether the unzipping should occur only for the 'top' zip file, or unzip 'all' recursively,
or only files 'in' the folder specified}
}
\description{
Unzip a zip file either at just the top level or recursively through the file
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
Claire Lunch \email{clunch@battelleecology.org}
}
