% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table_types.R
\name{update_table_types}
\alias{update_table_types}
\title{Update table types data frame in sysdata.rda}
\usage{
update_table_types()
}
\value{
A saved .rda file
}
\description{
Connect to the DPS database, pull in updated information about tables, and save locally.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
