% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddyStampCheck.R
\name{eddyStampCheck}
\alias{eddyStampCheck}
\title{Convert date stamps from character and check for only one record in a day}
\usage{
eddyStampCheck(tab)
}
\arguments{
\item{tab}{A table of SAE data}
}
\value{
The same table of SAE data, with time stamps converted and empty records representing a single day (filler records inserted during processing) removed.
}
\description{
Convert SAE time stamps to POSIX and check for missing data
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
