% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPI.R
\name{getAPI}
\alias{getAPI}
\title{Get the data from API}
\usage{
getAPI(apiURL, dpID = NA, token = NA)
}
\arguments{
\item{apiURL}{The API endpoint URL}

\item{dpID}{The identifier of the NEON data product to pull, in the form DPL.PRNUM.REV, e.g. DP1.10023.001}

\item{token}{User specific API token (generated within neon.datascience user accounts)}
}
\description{
Accesses the API with options to use the user-specific API token generated within neon.datascience user accounts.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Nate Mietkiewicz \email{mietkiewicz@battelleecology.org}
}
