% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileUrls.R
\name{getFileUrls}
\alias{getFileUrls}
\title{Get and store the file names, S3 URLs, file size, and download status (default = 0) in a data frame}
\usage{
getFileUrls(m.urls, include.provisional, token = NA)
}
\arguments{
\item{m.urls}{The monthly API URL for the AOP files}

\item{include.provisional}{T or F, should provisional data be included in downloaded files?}

\item{token}{User specific API token (generated within data.neonscience.org user accounts)}
}
\value{
A dataframe comprised of file names, S3 URLs, file size, and download status (default = 0)
}
\description{
Used to generate a data frame of available AOP files.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
