% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZipUrls.R
\name{getZipUrls}
\alias{getZipUrls}
\title{Get and store the file names, S3 URLs, file size, and download status (default = 0) in a data frame}
\usage{
getZipUrls(
  month.urls,
  avg,
  package,
  dpID,
  messages,
  tabl,
  token = NA_character_
)
}
\arguments{
\item{month.urls}{The monthly API URL for the URL files}

\item{avg}{Global variable for averaging interval}

\item{package}{Global varaible for package type (basic or expanded)}

\item{dpID}{Global variable for data product ID}

\item{messages}{Error/warning messages from previous steps}

\item{tabl}{Table name to get}

\item{token}{User specific API token (generated within neon.datascience user accounts)}
}
\value{
A dataframe comprised of file names, S3 URLs, file size, and download status (default = 0)
}
\description{
Used to generate a data frame of available zipfile URLs.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
Christine Laney \email{claney@battelleecology.org}
}
