% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackDataFilesArrow.R
\name{stackDataFilesArrow}
\alias{stackDataFilesArrow}
\title{Join data files in a unzipped NEON data package by table type}
\usage{
stackDataFilesArrow(folder, cloud.mode = FALSE, progress = TRUE, dpID)
}
\arguments{
\item{folder}{The location of the data}

\item{cloud.mode}{T or F, are data transferred from one cloud environment to another? If T, this function returns a list of url paths to data files.}

\item{progress}{T or F, should progress bars and messages be printed?}

\item{dpID}{The data product identifier}
}
\value{
One file for each table type is created and written.
}
\description{
Given a folder of unzipped files (unzipped NEON data file), do a full join of all data files, grouped by table type.
This should result in a small number of large files.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
