% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackDataFilesDuck.R
\name{stackDataFilesDuck}
\alias{stackDataFilesDuck}
\title{Align data files into a duckdb dataset}
\usage{
stackDataFilesDuck(urls, varset, tabl, package)
}
\arguments{
\item{urls}{The set of urls to be combined into a dataset.}

\item{varset}{A list of urls pointing to the set of variables files relevant to the url set.}

\item{tabl}{The table name of the table the url set represents.}

\item{package}{Basic or expanded data package?}
}
\value{
A duckdb dataset for the input data paths.
}
\description{
Takes a set of paths pointing to data files and a variables file and attempts to unify them into a duckdb dataset.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
