% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipsByURI.R
\name{zipsByURI}
\alias{zipsByURI}
\title{Get files from NEON ECS Bucket using URLs in stacked data function}
\usage{
zipsByURI(filepath, savepath = filepath, pick.files = TRUE,
  check.size = TRUE, unzip = TRUE)
}
\arguments{
\item{filepath}{The location of the zip file containing NEON data}

\item{savepath}{The location to save the output files from the ECS bucket}

\item{pick.files}{T or F, should the user be told the name of each file before downloading? Defaults to T. When working in batch mode, or other non-interactive workflow, use pick.files=F.}

\item{check.size}{T or F, should the user be told the total file size before downloading? Defaults to T. When working in batch mode, or other non-interactive workflow, use check.size=F.}

\item{unzip}{T or F, indicates if the downloaded zip files from ECS buckets should be unziped into the same directory. Defaults to T.}
}
\value{
A folder in the working directory (or in savepath, if specified), containing all zip files meeting query criteria.
}
\description{
Read in a set of URLs from stacked data and then download the data from the NEON ECS buckets
}
\examples{
\dontrun{
# To download stream morphology data from stacked data:
zipsByURI(filepath="~/filesToStack00131")
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin Cawley \email{kcawley@battelleecology.org}
}
